Change components location, dividing them in Home and Shared
