+++ src/app/layout.tsx
@@ -1,10 +1,11 @@
 import './globals.css'
-import Script from 'next/script'
 import type { Metadata } from 'next'
 
 import { config } from '@fortawesome/fontawesome-svg-core'
 import '@fortawesome/fontawesome-svg-core/styles.css'
 config.autoAddCss = false
+
+import { GTM } from '@/components/Shared/GTM'
 
 export const metadata: Metadata = {
   title: 'Hector Rubi | Portfolio',
@@ -18,20 +19,7 @@ export default function RootLayout({
   return (
     <html lang="en">
       <head>
-        {/* <!-- Google tag (gtag.js) --> */}
-        <Script
-          async
-          src="https://www.googletagmanager.com/gtag/js?id=G-RJFG3GMNZ1"
-        />
-        <Script id="google-analytics">
-          {`
-            window.dataLayer = window.dataLayer || [];
-            function gtag(){dataLayer.push(arguments);}
-            gtag('js', new Date());
-
-            gtag('config', 'G-RJFG3GMNZ1');
-          `}
-        </Script>
+        <GTM />
       </head>
       <body>{children}</body>
     </html>
