+++ src/components/Home/Hero/HeroPhrase/index.tsx
