+++ src/components/Home/Information/index.tsx
